error fix: see desc
when adding a new recipe and then trying to edit. it was pulling in the local state’s recipe instead of the props recipes meaning it wasn’t pulling in the latest recipe that had been added.

Now when adding a new recipe you can edit straight after and save.